+++ spec/hbw/features/tasks/task_counter_spec.rb
@@ -13,7 +13,7 @@ feature 'Check available tasks counter', js: true do
 
   describe 'when rendered' do
     it 'should contain assigned/unassigned tasks count' do
-      expect(tasks_count).to eq '36/3'
+      expect(tasks_count).to eq '37/3'
     end
   end
 
@@ -27,7 +27,7 @@ feature 'Check available tasks counter', js: true do
     end
 
     it 'should contain two tabs with list' do
-      expect(page).to have_content 'My tasks (36)'
+      expect(page).to have_content 'My tasks (37)'
       expect(page).to have_content 'Unclaimed tasks (3)'
 
       expect(popup_tasks_list_content).to eq(
@@ -57,7 +57,7 @@ feature 'Check available tasks counter', js: true do
 
       click_on_tasks_counter
 
-      expect(page).not_to have_content 'My tasks (36)'
+      expect(page).not_to have_content 'My tasks (37)'
       expect(page).not_to have_content 'Unclaimed tasks (3)'
 
       expect_widget_presence