+++ spec/hbw/features/tasks/task_list_spec.rb
@@ -15,7 +15,7 @@ feature 'Check table with tasks', js: true do
     click_on 'Tasks'
     expect_widget_presence
     expect(page).to have_content 'Open tasks'
-    expect(page).to have_content 'My tasks (36)'
+    expect(page).to have_content 'My tasks (37)'
     expect(page).to have_content 'Unclaimed tasks (3)'
   end
 
@@ -28,7 +28,7 @@ feature 'Check table with tasks', js: true do
           ['Medium', 'Assigned task', ' Test name', '—', "expired (#{years_since(first_task_due_date)}y past due date)"],
           ['High', 'Other assigned task', ' Test name', 'Some test description', "expired (#{years_since(second_task_due_date)}y past due date)"],
           ['High', "Enter customer's address", ' New Customer', '—', '30 Jun 2016'],
-          *Array.new(33) { ['High', 'Check test form', ' Test name', '—', '30 Jun 2016'] }
+          *Array.new(34) { ['High', 'Check test form', ' Test name', '—', '30 Jun 2016'] }
         ]
       )
     end
@@ -54,7 +54,7 @@ feature 'Check table with tasks', js: true do
     end
 
     it 'for assigned task' do
-      expect(tasks_table_content.length).to eq(36)
+      expect(tasks_table_content.length).to eq(37)
 
       click_on_task_table_row(1)