+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v2.6.2 [unreleased]
 ### Features
 - [#556](https://github.com/latera/homs/pull/556) Add css variable for links color.
 - [#557](https://github.com/latera/homs/pull/557) Customize static fonts.
+- [#558](https://github.com/latera/homs/pull/558) Add dynamic `disable_if` and `delete_if` options for radio button.
 
 v2.6.1 [2021-05-17]
 -------------------